added installer v1.0.0

+++ install.ps1
@@ -1,6 +1,7 @@
 cls
-rm ./updateAndRun.ps1
-rm ./.git -r -fo
+if(Test-Path -Path './.git' -PathType Container ){
+    rm ./.git -r -fo
+}
 git init
 git pull --set-upstream http://github.com/akurai/http_practice main
 git branch -m master main
